+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Upgrade query builder JSON serializer to Jackson 2.8.0
 * Add geospatial post filtering (ensures accuracy with any tree levels)
+* Set default number of indexing threads to number of processors available to the JVM
 * Fix mapping of timestamps and dates by their underlying numeric value (#177)
 Merged from 2.2.7.1:
 * Fix mapper referenced by alias in sortFields